Implement basic gameplay persistence (autosave)
This implements a rudimentary autosave/autoload feature. When the game is
closed, the state and its metadata are saved to the hard drive. When the
game is played again, the state is automatically loaded.

This is called "gameplay persistence" because the game resumes where it was
last left off, with no explicit saving/loading action by the user.

Additionally, to guard against power loss, the game is autosaved every 10
seconds.

Savestates and metadata are stored alongside the game by replacing the file
extension.

If a different game client is used, Kodi will warn the user that it will
overwrite their existing save. This is because savestates are incompatible
between emulators, and storing more than a single savestate isn't
implemented yet.