[addons] change OnEnabled and OnDisabled to use AddonPtr
Before was the call done with the ID string only and limited to PVR
addon.

This change them to use the on CAddonManager created IAddon class
to give related addon.

Test here with PVR and Peripheral was successful.

This is required part on next changes where the CBinaryAddonManager
handle the addon changes and to inform related parts thread safe.

As note, PVR and some other binary addons can not work with the
"AddonEvents"! There must be e.g. before uninstall confirmed
that the running addon is stopped and by install thats present.

Further has I seen that currently e.g. on peripheral addons the disable
and enable not work with "AddonEvents" from the above described reasons.