[addons][inputstream] separate "C" and "C++" header parts in own files
This is done to confirm a pure "C" within interface and to allow
use for other languages where need a "C" as base.

This changes are the initial part, there still few included headers
where use C++!

This becomes improved on following changes, for here to have the
other commits more clean and see if relate to "C" between Kodi
or to "C++" on addon.